Move server.go into pkg/provider/server
