+++ cmd/gmsa-account-provider/main.go
@@ -78,7 +78,7 @@ func (a *GMSAAccountProvider) Run(cmd *cobra.Command, _ []string) error {
 		return err
 	}
 
-	server := provider.HTTPServer{
+	server := server.HTTPServer{
 		Handler: server.NewHandler(getter.Namespaced[*v1.Secret](secrets, a.Namespace)),
 	}
